[Refactor] #244 AuthService에서 서비스 간 의존성 제거하고 facade 생성
