--- src/main/java/org/sopt/kareer/domain/member/service/MemberService.java
@@ -1,286 +1,0 @@
-package org.sopt.kareer.domain.member.service;
-
-import lombok.RequiredArgsConstructor;
-import org.sopt.kareer.domain.member.dto.request.MemberOnboardV2Request;
-import org.sopt.kareer.domain.member.dto.request.MemberTermsRequest;
-import org.sopt.kareer.domain.member.dto.response.*;
-import org.sopt.kareer.domain.member.entity.Member;
-import org.sopt.kareer.domain.member.entity.MemberTerm;
-import org.sopt.kareer.domain.member.entity.MemberVisa;
-import org.sopt.kareer.domain.member.entity.enums.LocalizedOnboardCategoryType;
-import org.sopt.kareer.domain.member.entity.enums.MemberStatus;
-import org.sopt.kareer.domain.member.exception.MemberErrorCode;
-import org.sopt.kareer.domain.member.exception.MemberException;
-import org.sopt.kareer.domain.member.repository.MemberRepository;
-import org.sopt.kareer.domain.member.repository.MemberTermRepository;
-import org.sopt.kareer.domain.member.repository.MemberVisaRepository;
-import org.sopt.kareer.domain.member.service.dto.request.MypageCommand;
-import org.sopt.kareer.domain.member.util.PassportOcrParser;
-import org.sopt.kareer.domain.member.util.VisaOcrParser;
-import org.sopt.kareer.domain.term.entity.Term;
-import org.sopt.kareer.domain.term.exception.TermErrorCode;
-import org.sopt.kareer.domain.term.exception.TermException;
-import org.sopt.kareer.domain.term.service.TermService;
-import org.sopt.kareer.global.document.exception.DocumentErrorCode;
-import org.sopt.kareer.global.document.exception.DocumentException;
-import org.sopt.kareer.global.document.service.DocumentProcessingService;
-import org.sopt.kareer.global.exception.customexception.GlobalException;
-import org.sopt.kareer.global.exception.errorcode.GlobalErrorCode;
-import org.sopt.kareer.global.oauth.dto.OAuthAttributes;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class MemberService {
-
-    private final MemberRepository memberRepository;
-    private final MemberVisaRepository memberVisaRepository;
-    private final TermService termService;
-    private final MemberTermRepository memberTermRepository;
-    private final MemberDeletionService memberDeletionService;
-    private final LocalizedOnboardQueryService localizedOnboardQueryService;
-    private final DocumentProcessingService documentProcessingService;
-    private final VisaOcrParser visaOcrParser;
-    private final PassportOcrParser passportOcrParser;
-
-    public Member getById(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-    }
-
-    @Transactional
-    public Member findOrCreateByOAuth(OAuthAttributes attributes) {
-        return memberRepository.findByProviderAndProviderId(attributes.provider(), attributes.providerId())
-                .orElseGet(() -> createNewMember(attributes));
-    }
-
-    private Member createNewMember(OAuthAttributes attributes) {
-        Member member = Member.createOAuthMember(
-                attributes.name(),
-                attributes.provider(),
-                attributes.providerId(),
-                attributes.picture(),
-                attributes.email()
-        );
-        try {
-            return memberRepository.save(member);
-        } catch (DataIntegrityViolationException ex) {
-            return memberRepository.findByProviderAndProviderId(attributes.provider(), attributes.providerId())
-                    .orElseThrow(() -> new GlobalException(GlobalErrorCode.INTERNAL_SERVER_ERROR));
-        }
-    }
-
-
-    public MemberInfoResponse getMemberInfo(Long memberId) {
-        Member member = getById(memberId);
-        MemberVisa memberVisa = memberVisaRepository.findActiveByMemberId(memberId)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.VISA_NOT_FOUND));
-        String countryLabel = localizedOnboardQueryService.resolveLabelByCode(LocalizedOnboardCategoryType.COUNTRY,
-                member.getCountryCode());
-        String primaryMajorLabel = localizedOnboardQueryService.resolveLabelByCode(LocalizedOnboardCategoryType.MAJOR,
-                member.getPrimaryMajorCode());
-        String universityLabel = localizedOnboardQueryService.resolveLabelByCode(
-                LocalizedOnboardCategoryType.UNIVERSITY, member.getUniversityCode());
-        String degreeLabel = localizedOnboardQueryService.resolveLabelByCode(LocalizedOnboardCategoryType.DEGREE,
-                member.getDegreeCode());
-        String englishLevelLabel = localizedOnboardQueryService.resolveLabelByCode(
-                LocalizedOnboardCategoryType.ENGLISH_LEVEL, member.getEnglishLevelCode());
-
-        return MemberInfoResponse.of(
-                member,
-                memberVisa,
-                countryLabel,
-                primaryMajorLabel,
-                universityLabel,
-                degreeLabel,
-                englishLevelLabel
-        );
-    }
-
-    @Transactional
-    public void onboardMemberV2(MemberOnboardV2Request request, Long memberId) {
-        Member member = getById(memberId);
-        String fieldOfInterest = String.join(",", request.fieldsOfInterests());
-        String preparationStatus = String.join(",", request.preparationStatuses());
-
-        member.updateInfoV2(
-                request.name(),
-                request.birthDate(),
-                request.countryCode(),
-                request.universityCode(),
-                request.englishLevel().getDescription(),
-                fieldOfInterest,
-                preparationStatus,
-                request.languageLevel(),
-                request.degree().getDescription(),
-                request.expectedGraduationDate(),
-                request.primaryMajorCode(),
-                request.secondaryMajor(),
-                request.targetJob(),
-                request.targetJobSkill(),
-                request.personalBackground()
-        );
-
-        MemberVisa memberVisa = MemberVisa.createMemberVisa(
-                member,
-                request.visaType(),
-                request.visaExpiredAt(),
-                request.visaStartDate()
-        );
-        memberVisaRepository.save(memberVisa);
-    }
-
-    public MemberStatusResponse getMemberStatus(Long memberId) {
-        Member member = getById(memberId);
-
-        MemberVisa memberVisa = memberVisaRepository.findActiveByMemberId(memberId)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.VISA_NOT_FOUND));
-
-        return MemberStatusResponse.from(member, memberVisa);
-    }
-
-    public MypageResponse getMypage(Long memberId) {
-        Member member = getById(memberId);
-        MemberVisa memberVisa = memberVisaRepository.findActiveByMemberId(memberId)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.VISA_NOT_FOUND));
-        String countryLabel = localizedOnboardQueryService.resolveLabelByCode(LocalizedOnboardCategoryType.COUNTRY,
-                member.getCountryCode());
-        String primaryMajorLabel = localizedOnboardQueryService.resolveLabelByCode(LocalizedOnboardCategoryType.MAJOR,
-                member.getPrimaryMajorCode());
-        String universityLabel = localizedOnboardQueryService.resolveLabelByCode(
-                LocalizedOnboardCategoryType.UNIVERSITY, member.getUniversityCode());
-        String degreeLabel = localizedOnboardQueryService.resolveLabelByCode(LocalizedOnboardCategoryType.DEGREE,
-                member.getDegreeCode());
-        String englishLevelLabel = localizedOnboardQueryService.resolveLabelByCode(
-                LocalizedOnboardCategoryType.ENGLISH_LEVEL, member.getEnglishLevelCode());
-
-        return MypageResponse.of(member, memberVisa, countryLabel, primaryMajorLabel, universityLabel, degreeLabel,
-                englishLevelLabel);
-
-    }
-
-    @Transactional
-    public void updateMypage(Long memberId, MypageCommand command) {
-        Member member = getById(memberId);
-        MemberVisa memberVisa = memberVisaRepository.findActiveByMemberId(memberId)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.VISA_NOT_FOUND));
-
-        member.updateProfile(
-                command.targetJob(),
-                command.birthDate(),
-                command.countryCode(),
-                command.degree(),
-                command.universityCode(),
-                command.primaryMajorCode(),
-                command.secondaryMajor(),
-                command.languageLevel(),
-                command.englishLevel()
-        );
-        memberVisa.updateVisa(command.visaType(), command.visaExpiredAt());
-
-    }
-
-    @Transactional
-    public void deleteMember(Long memberId) {
-        Member member = getById(memberId);
-        memberDeletionService.deleteMember(member);
-    }
-
-
-    public OcrVisaResponse getVisaOcr(MultipartFile file) {
-        try {
-            String text = documentProcessingService.extractText(file);
-            VisaOcrParser.VisaInfo visaInfo = visaOcrParser.parse(text);
-
-            return OcrVisaResponse.from(visaInfo);
-        } catch (DocumentException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new DocumentException(
-                    DocumentErrorCode.OCR_PROCESSING_FAILED
-            );
-        }
-    }
-
-    public OcrPassportResponse getPassportOcr(MultipartFile file) {
-        try {
-            String text = documentProcessingService.extractText(file);
-            PassportOcrParser.PassportInfo passportInfo = passportOcrParser.parse(text);
-
-            return OcrPassportResponse.from(passportInfo);
-        } catch (DocumentException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new DocumentException(
-                    DocumentErrorCode.OCR_PROCESSING_FAILED
-            );
-        }
-    }
-
-    @Transactional
-    public void agreeTerms(Long memberId, MemberTermsRequest request) {
-        if (memberTermRepository.existsByMemberId(memberId)) {
-            throw new TermException(TermErrorCode.ALREADY_AGREED_TERMS);
-        }
-
-        Member member = getById(memberId);
-        List<MemberTermsRequest.TermAgreement> agreements = request.agreements();
-
-        Map<Long, Boolean> agreementMap = agreements.stream()
-                .collect(Collectors.toMap(
-                        MemberTermsRequest.TermAgreement::termId,
-                        MemberTermsRequest.TermAgreement::agreed,
-
-                        // 중복된 term이 있는지 체크
-                        (existing, replacement) -> {
-                            throw new TermException(TermErrorCode.DUPLICATE_TERM);
-                        }
-                ));
-
-        // 받아야하는 약관들 리스트
-        List<Term> activeTerms = termService.getActiveTerms();
-
-        // 필요한 약관들에 대해 잘 요청됐는지 검증
-        Set<Long> activeTermIds = activeTerms.stream()
-                .map(Term::getId)
-                .collect(Collectors.toSet());
-
-        if (!activeTermIds.equals(agreementMap.keySet())) {
-            throw new TermException(TermErrorCode.MISSING_TERM);
-        }
-
-        List<MemberTerm> memberTerms = activeTerms.stream()
-                .map(term -> {
-                    Boolean agreed = agreementMap.get(term.getId());
-
-                    // 필수 약관인데 동의하지 않은 경우
-                    if (term.isRequired() && !Boolean.TRUE.equals(agreed)) {
-                        throw new TermException(TermErrorCode.REQUIRED_TERM_NOT_AGREED);
-                    }
-
-                    return MemberTerm.create(agreed, member, term);
-                })
-                .toList();
-
-        memberTermRepository.saveAll(memberTerms);
-    }
-
-
-    public MemberCompletionResponse getCompletion(Long memberId){
-        Member member = getById(memberId);
-
-        boolean onboardingRequired = member.getStatus().equals(MemberStatus.PENDING);
-        boolean agreedTerm = memberTermRepository.existsByMemberIdAndAgreedTrue(memberId);
-
-        return MemberCompletionResponse.of(onboardingRequired, agreedTerm);
-    }
-}
