+++ src/main/java/org/sopt/kareer/global/auth/controller/AuthController.java
@@ -11,7 +11,7 @@ import org.sopt.kareer.global.config.swagger.SwaggerResponseDescription;
 import org.sopt.kareer.global.auth.dto.request.TokenExchangeRequest;
 import org.sopt.kareer.global.auth.dto.response.TokenExchangeResponse;
 import org.sopt.kareer.global.auth.dto.response.TokenResponse;
-import org.sopt.kareer.global.auth.service.AuthService;
+import org.sopt.kareer.global.auth.facade.AuthFacade;
 import org.sopt.kareer.global.response.BaseResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,14 +26,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/auth")
 public class AuthController {
 
-    private final AuthService authService;
+    private final AuthFacade authFacade;
 
     @PostMapping("/code/exchange")
     @Operation(summary = "로그인 코드 교환", description = "OAuth 로그인 후 발급된 임시 코드를 AT/RT로 교환합니다.")
     @CustomExceptionDescription(SwaggerResponseDescription.AUTH_LOGIN_CODE)
     public ResponseEntity<BaseResponse<TokenExchangeResponse>> exchange(@Valid @RequestBody TokenExchangeRequest request,
                                                                         HttpServletResponse response) {
-        TokenExchangeResponse exchangeResponse = authService.exchange(request, response);
+        TokenExchangeResponse exchangeResponse = authFacade.exchange(request, response);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(BaseResponse.ok(exchangeResponse, "로그인 코드가 교환되었습니다."));
     }
@@ -43,7 +43,7 @@ public class AuthController {
     @CustomExceptionDescription(SwaggerResponseDescription.AUTH_REISSUE)
     public ResponseEntity<BaseResponse<TokenResponse>> reissue(HttpServletRequest request,
                                                                HttpServletResponse response) {
-        TokenResponse tokenResponse = authService.reissue(request, response);
+        TokenResponse tokenResponse = authFacade.reissue(request, response);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(BaseResponse.ok(tokenResponse, "토큰이 재발급되었습니다."));
     }
@@ -53,7 +53,7 @@ public class AuthController {
     @CustomExceptionDescription(SwaggerResponseDescription.AUTH_SIGN_OUT)
     public ResponseEntity<BaseResponse<Void>> signOut(HttpServletRequest request,
                                                       HttpServletResponse response) {
-        authService.signOut(request, response);
+        authFacade.signOut(request, response);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(BaseResponse.ok("로그아웃되었습니다."));
     }
